New: examples Update: README.md
