--- example/linear_runnable_test.go
@@ -1,17 +1,0 @@
-package example
-
-import (
-	"fmt"
-	"math/rand"
-)
-
-func ExamplePerm() {
-	for _, value := range rand.Perm(5) {
-		fmt.Println(value)
-	}
-	// Unordered output: 4
-	// 2
-	// 1
-	// 3
-	// 0
-}
